disease_agent/requirements.txt
torch
transformers
sentence-transformers
faiss-gpu
scikit-learn
requests
python-dotenv